Refactor baseController into separate file
baseController code is the majority of controller.go. Split baseController code
into base.go.

Subsequent commit will move config to be stored in State Store. baseController
will have a copy of config that should be handled differently from structure's
access to config.

Splitting out baseController will make it easier to handle different usages of
config
